requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
influxdb-client==1.38.0
scylladb-driver==3.28.0
python-dotenv==1.0.0
pydantic==2.5.2
apscheduler==3.10.4
numpy==1.26.2
